Crate tvm_frontend

TVM is a compiler stack for deep learning systems.

This crate provides an idiomatic Rust API for TVM runtime frontend.

One particular use case is that given optimized deep learning model artifacts, already compiled with TVM, which include a shared library lib.so, graph.json and a byte-array param.params, one can load them in Rust to create a graph runtime. Then, run the model for some inputs and get the desired predictions all in Rust.

Checkout the examples repository for more details.

Re-exports

pub use bytearray::TVMByteArray;

pub use context::TVMContext;

pub use context::TVMDeviceType;

pub use errors::*;

pub use function::Function;

pub use module::Module;

pub use ndarray::empty;

pub use ndarray::NDArray;

pub use ty::TVMType;

pub use value::TVMArgValue;

pub use value::TVMRetValue;

Modules

bytearray
Provides TVMByteArray used for passing the model parameters (stored as byte-array) to a runtime module.

context
Provides TVMContext and related device specific queries.

errors
This module implements TVM custom Error, ErrorKind and Result types.

function
This module provides an idiomatic Rust API for creating and working with TVM functions.

module
Provides the Module type and methods for working with runtime TVM modules.

ndarray
This module implements the NDArray type for working with TVM tensors or coverting from a Rust's ndarray to TVM NDArray.

ty
This module implements the required conversions from Rust types to TVM types.

value
This module implements TVMArgValue and TVMRetValue types and their conversions to other supported TVMValues. TVMRetValue is the owned version of TVMArgValue.

Macros

call_packed
Convenient macro for calling TVM packed functions by providing a function identifier and some arguments. This macro outputs a Result type and let user to perform proper error handling.

register_global_func
Convenient macro for registering functions from frontend to backend as global TVM packed functions without overriding. If overriding an existing function is needed use the function::register function instead.

Functions

get_last_error
Gets the last error message

version
Outputs the current TVM version