Macro tvm_frontend::register_global_func

macro_rules! register_global_func {
    {
        $(#[$m:meta])*
        fn $fn_name:ident($args:ident : &[TVMArgValue]) -> Result<TVMRetValue> {
            $($code:tt)*
        }
    } => { ... };
}

Convenient macro for registering functions from frontend to backend as global TVM packed functions without overriding. If overriding an existing function is needed use the function::register function instead.

Example

register_global_func! {
    fn sum(args: &[TVMArgValue]) -> Result<TVMRetValue> {
        let mut ret = 0f64;
        for arg in args.iter() {
            ret += arg.to_float();
        }
        let ret_val = TVMRetValue::from(&ret);
        Ok(ret_val)
    }
}

let mut registered = function::Builder::default();
registered.get_function("sum", true);
assert!(registered.func.is_some());
registered.args(&[10f64, 20f64, 30f64]);
assert_eq!(registered.invoke().unwrap().to_float(), 60f64);