Struct tvm_frontend::ndarray::NDArray

pub struct NDArray { /* fields omitted */ }

See the module-level documentation for more details.

Wrapper around TVM array handle.

Methods

impl NDArray

pub fn handle(&self) -> TVMArrayHandle

Returns the underlying array handle.

pub fn is_view(&self) -> bool

pub fn shape(&self) -> Option<&mut [usize]>

Returns the shape of the NDArray.

pub fn size(&self) -> Option<usize>

Returns the total number of entries of the NDArray.

pub fn ctx(&self) -> TVMContext

Returns the context which the NDArray was defined.

pub fn dtype(&self) -> TVMType

Returns the type of the entries of the NDArray.

pub fn ndim(&self) -> usize

Returns the number of dimensions of the NDArray.

pub fn strides(&self) -> Option<&[usize]>

Returns the strides of the underlying NDArray.

pub fn is_contiguous(&self) -> Result<bool>

pub fn byte_offset(&self) -> isize

pub fn to_vec<T>(&self) -> Result<Vec<T>>

Flattens the NDArray to a Vec of the same type in cpu.

Example

let shape = &mut [4];
let mut data = vec![1i32, 2, 3, 4];
let ctx = TVMContext::cpu(0);
let mut ndarray = empty(shape, ctx, TVMType::from("int"));
ndarray.copy_from_buffer(&mut data);
assert_eq!(ndarray.shape(), Some(shape));
assert_eq!(ndarray.to_vec::<i32>().unwrap(), data);

pub fn to_bytearray(&self) -> Result<TVMByteArray>

Converts the NDArray to TVMByteArray.

pub fn copy_from_buffer<T: Num32>(&mut self, data: &mut [T])

Creates an NDArray from a mutable buffer of types i32, u32 or f32 in cpu.

Example

let shape = &mut [2];
let mut data = vec![1f32, 2];
let ctx = TVMContext::gpu(0);
let mut ndarray = empty(shape, ctx, TVMType::from("int"));
ndarray.copy_from_buffer(&mut data);

Note: if something goes wrong during the copy, it will panic from TVM side. See TVMArrayCopyFromBytes in c_runtime_api.h.

pub fn copy_to_ndarray(&self, target: NDArray) -> Result<NDArray>

Copies the NDArray to another target NDArray.

pub fn copy_to_ctx(&self, target: &TVMContext) -> Result<NDArray>

Copies the NDArray to a target context.

pub fn from_rust_ndarray<T: Num32 + Copy>(
rnd: &ArrayD<T>,
ctx: TVMContext,
dtype: TVMType
) -> Result<Self>

Converts a Rust's ndarray to TVM NDArray.